requirements.txt
Pillow ==8.3.2
Django ==3.2.7
djangorestframework ==3.12.4
whitenoise== 5.3.0
djangorestframework-simplejwt ==4.8.0